query score fetch completion

+++ src/main/java/io/credable/loanmanagementsystem/data/dto/ScoringDTO.java
@@ -13,7 +13,7 @@ public class ScoringDTO {
     int  id;
     String CustomerNumber;
     int score;
-    int limitAmount;
+    Double limitAmount;
     String exclusion;
     String exclusionReason;
